Feat: playlist array
