+++ package.json
@@ -9,20 +9,19 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@googleapis/youtube": "^14.0.0",
+    "@supabase/supabase-js": "^2.39.8",
     "@tanstack/react-query": "^5.28.4",
-    "axios": "^1.6.8",
-    "google-api-python-client": "^0.0.1",
-    "next": "14.1.3",
+    "jotai": "^2.7.1",
+    "next": "^14.1.3",
     "react": "^18",
-    "react-dom": "^18",
-    "react-router-dom": "^6.22.3"
+    "react-dom": "^18"
   },
   "devDependencies": {
     "@types/node": "^20",
     "@types/react": "^18",
     "@types/react-dom": "^18",
     "autoprefixer": "^10.0.1",
+    "daisyui": "^4.7.3",
     "eslint": "^8",
     "eslint-config-next": "14.1.3",
     "postcss": "^8",
