+++ src/app/devCookiePlaylist/[id]/page.tsx
@@ -1,4 +1,4 @@
-import PlaylistSearchList from "@/devCookiePlaylistComponents/PlaylistSearchList";
+import PlaylistSearchList from "@/app/devCookiePlaylistComponents/PlaylistSearchList";
 import React from "react";
 
 function page() {
